Remove "row" class when displaying history
The "row" class adds some negative margin left and right, which breaks
the alignment when the history is not rendered in a modal window.
The default behavior to this point was to display the history in a modal
window. With this work, the history will also be available on its own.
Therefore some adjustments in the rendering is needed.